+++ tests/test_deprecated.py
@@ -1790,25 +1790,25 @@ def test_dep_ecliptic():
         check_dep(galsim.DMS_Angle,'66:33:38.55'))
     el, b = north_pole.ecliptic()
     # North pole should have b=90 degrees, with el being completely arbitrary.
-    np.testing.assert_almost_equal(b.rad(), pi/2, decimal=6)
+    np.testing.assert_almost_equal(check_dep(b.rad), pi/2, decimal=6)
 
     south_pole = galsim.CelestialCoord(
         check_dep(galsim.HMS_Angle,'06:00:00.00'),
         check_dep(galsim.DMS_Angle,'-66:33:38.55'))
     el, b = south_pole.ecliptic()
     # South pole should have b=-90 degrees, with el being completely arbitrary.
-    np.testing.assert_almost_equal(b.rad(), -pi/2, decimal=6)
+    np.testing.assert_almost_equal(check_dep(b.rad), -pi/2, decimal=6)
 
     # Also confirm that positions that should be the same in equatorial and ecliptic coordinates are
     # actually the same:
     vernal_equinox = galsim.CelestialCoord(0.*galsim.radians, 0.*galsim.radians)
     el, b = vernal_equinox.ecliptic()
-    np.testing.assert_almost_equal(b.rad(), 0., decimal=6)
-    np.testing.assert_almost_equal(el.rad(), 0., decimal=6)
+    np.testing.assert_almost_equal(check_dep(b.rad), 0., decimal=6)
+    np.testing.assert_almost_equal(check_dep(el.rad), 0., decimal=6)
     autumnal_equinox = galsim.CelestialCoord(pi*galsim.radians, 0.*galsim.radians)
     el, b = autumnal_equinox.ecliptic()
-    np.testing.assert_almost_equal(el.rad(), pi, decimal=6)
-    np.testing.assert_almost_equal(b.rad(), 0., decimal=6)
+    np.testing.assert_almost_equal(check_dep(el.rad), pi, decimal=6)
+    np.testing.assert_almost_equal(check_dep(b.rad), 0., decimal=6)
 
     # Finally, test the results of using a date to get ecliptic coordinates with respect to the sun,
     # instead of absolute ones. For this, use dates and times of vernal and autumnal equinox
@@ -1821,46 +1821,50 @@ def test_dep_ecliptic():
     el, b = vernal_equinox.ecliptic(epoch=2014)
     el_rel, b_rel = vernal_equinox.ecliptic(epoch=2014, date=vernal_eq_date)
     # Vernal equinox: should have (el, b) = (el_rel, b_rel) = 0.0
-    np.testing.assert_almost_equal(el_rel.rad(), el.rad(), decimal=3)
-    np.testing.assert_almost_equal(b_rel.rad(), b.rad(), decimal=6)
+    np.testing.assert_almost_equal(check_dep(el_rel.rad), check_dep(el.rad), decimal=3)
+    np.testing.assert_almost_equal(check_dep(b_rel.rad), check_dep(b.rad), decimal=6)
     # Now do the autumnal equinox: should have (el, b) = (pi, 0) = (el_rel, b_rel) when we look at
     # the time of the vernal equinox.
     el, b = autumnal_equinox.ecliptic(epoch=2014)
     el_rel, b_rel = autumnal_equinox.ecliptic(epoch=2014, date=vernal_eq_date)
-    np.testing.assert_almost_equal(el_rel.rad(), el.rad(), decimal=3)
-    np.testing.assert_almost_equal(b_rel.rad(), b.rad(), decimal=6)
+    np.testing.assert_almost_equal(check_dep(el_rel.rad), check_dep(el.rad), decimal=3)
+    np.testing.assert_almost_equal(check_dep(b_rel.rad), check_dep(b.rad), decimal=6)
     # And check that if it's the date of the autumnal equinox (sun at (180, 0)) but we're looking at
     # the position of the vernal equinox (0, 0), then (el_rel, b_rel) = (-180, 0)
     autumnal_eq_date = datetime.datetime(2014,9,23,2,29,0)
     el_rel, b_rel = vernal_equinox.ecliptic(epoch=2014, date=autumnal_eq_date)
-    np.testing.assert_almost_equal(el_rel.rad(), -pi, decimal=3)
-    np.testing.assert_almost_equal(b_rel.rad(), 0., decimal=6)
+    np.testing.assert_almost_equal(check_dep(el_rel.rad), -pi, decimal=3)
+    np.testing.assert_almost_equal(check_dep(b_rel.rad), 0., decimal=6)
     # And check that if it's the date of the vernal equinox (sun at (0, 0)) but we're looking at
     # the position of the autumnal equinox (180, 0), then (el_rel, b_rel) = (180, 0)
     el_rel, b_rel = autumnal_equinox.ecliptic(epoch=2014, date=vernal_eq_date)
-    np.testing.assert_almost_equal(el_rel.rad(), pi, decimal=3)
-    np.testing.assert_almost_equal(b_rel.rad(), 0., decimal=6)
+    np.testing.assert_almost_equal(check_dep(el_rel.rad), pi, decimal=3)
+    np.testing.assert_almost_equal(check_dep(b_rel.rad), 0., decimal=6)
 
     # Check round-trips: go from CelestialCoord to ecliptic back to equatorial, and make sure
     # results are the same.  This includes use of a function that isn't available to users, but we
     # use it for a few things so we should still make sure it's working properly.
     from galsim.celestial import _ecliptic_to_equatorial
     north_pole_2 = _ecliptic_to_equatorial(north_pole.ecliptic(epoch=2014), 2014)
-    np.testing.assert_almost_equal(north_pole.ra.rad(), north_pole_2.ra.rad(), decimal=6)
-    np.testing.assert_almost_equal(north_pole.dec.rad(), north_pole_2.dec.rad(), decimal=6)
+    np.testing.assert_almost_equal(check_dep(north_pole.ra.rad),
+                                   check_dep(north_pole_2.ra.rad), decimal=6)
+    np.testing.assert_almost_equal(check_dep(north_pole.dec.rad),
+                                   check_dep(north_pole_2.dec.rad), decimal=6)
     south_pole_2 = _ecliptic_to_equatorial(south_pole.ecliptic(epoch=2014), 2014)
-    np.testing.assert_almost_equal(south_pole.ra.rad(), south_pole_2.ra.rad(), decimal=6)
-    np.testing.assert_almost_equal(south_pole.dec.rad(), south_pole_2.dec.rad(), decimal=6)
+    np.testing.assert_almost_equal(check_dep(south_pole.ra.rad),
+                                   check_dep(south_pole_2.ra.rad), decimal=6)
+    np.testing.assert_almost_equal(check_dep(south_pole.dec.rad),
+                                   check_dep(south_pole_2.dec.rad), decimal=6)
     vernal_equinox_2 = _ecliptic_to_equatorial(vernal_equinox.ecliptic(epoch=2014), 2014)
-    np.testing.assert_almost_equal(vernal_equinox.ra.rad(), vernal_equinox_2.ra.rad(), decimal=6)
-    np.testing.assert_almost_equal(vernal_equinox.dec.rad(), vernal_equinox_2.dec.rad(),
-                                      decimal=6)
+    np.testing.assert_almost_equal(check_dep(vernal_equinox.ra.rad),
+                                   check_dep(vernal_equinox_2.ra.rad), decimal=6)
+    np.testing.assert_almost_equal(check_dep(vernal_equinox.dec.rad),
+                                   check_dep(vernal_equinox_2.dec.rad), decimal=6)
     autumnal_equinox_2 = _ecliptic_to_equatorial(autumnal_equinox.ecliptic(epoch=2014), 2014)
-    np.testing.assert_almost_equal(autumnal_equinox.ra.rad(), autumnal_equinox_2.ra.rad(),
-                                      decimal=6)
-    np.testing.assert_almost_equal(autumnal_equinox.dec.rad(), autumnal_equinox_2.dec.rad(),
-                                      decimal=6)
-
+    np.testing.assert_almost_equal(check_dep(autumnal_equinox.ra.rad),
+                                   check_dep(autumnal_equinox_2.ra.rad), decimal=6)
+    np.testing.assert_almost_equal(check_dep(autumnal_equinox.dec.rad),
+                                   check_dep(autumnal_equinox_2.dec.rad), decimal=6)
 
 if __name__ == "__main__":
     test_dep_bandpass()